Fix subfolder url rewritting for external domains
The problematic behaviour is that URLs that are included as content but
belong to a different domain are not excluded from subdirectories url
rewritting.
The fix introduces a check in the main function that does the url re-
writting (function reverse_hard_link, in
transifex-live-integration-rewrite.php) that validates the host part
of the url in every link processed to guarantee that we leave unchanged
URLs in different domains.