Support WP install in subdirectory
We add a new option in the plugin settings in order to support
installation of WP under a subdirectory. So if the root folder of the
webserver hosting WP install is '/var/www/html' and is accessible via
'https://my.domain.com/' and if the WP is installed in the sub-
directory 'var/www/html/cms/' then the plugin option should be
'cms' and WP will be accessible via 'https://my.domain/com/cms/'.

Note: we support one level of subdirectory, not nested subdirs.

This new option is
compatible with all of the plugin's SEO settings (ie 'Disabled',
'Subdirectory' or 'Subdomain').