Support WP install in subdirectory
We add a new option in the plugin settings in order to support
installation of WP under a subdirectory. So if the root folder of the
webserver hosting WP install is '/var/www/html' and is accessible via
'https://my.domain.com/' and if the WP is installed in the sub-
directory '/var/www/html/cms/' then WP admin UI will be accessible
via 'https://my.domain/com/cms/wp-admin/...' and WP content from
'https://my.domain/com/...'.

This new option is
compatible with all of the plugin's SEO settings (ie 'Disabled',
'Subdirectory' or 'Subdomain').