Fix hreflang tags for subdirectory URL option
When the plug-in is configured to create localized URL as
subdirectories i.e.:

  https://www.xxx.com/enact-sample-post/...

if the URL path coincide to start from one of the used language codes
(in the current example enact-sample-post/... starts from en - the
source language code) the existing code stripped off the presumed
lang code which resulted in wrong hreflang tags. In the given example
the calculated hreflang tag was:

  <link rel="alternate" href="www.xxx.com/act-sample-post/..."
   hreflang="en>

This commit fixes this wrong hrelang tags generation.

We also use an updated Ubuntu image for github actions i.e. changes
in `.github/workflows/main.yml`